+++ src/lib.rs
@@ -67,6 +67,7 @@ pub use serde_impls::serde_bid;
 #[cfg(any(feature = "trig", feature = "exp-log"))]
 pub use math::{e, ln10, ln2, pi};
 
+pub use classify::IeeeClass;
 pub use decimal::{Decimal128, Decimal128BuildError};
 pub use status::{RoundingMode, Status};